# application/static/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A RESTful web service for managing payments via the Worldpay processing platform.",
    "version": "v1",
    "title": "Payment Gateway"
  },
  "basePath": "/payment-gateway/api/v1",
  "tags": [
    {
      "name": "payments",
      "description": "Operations for managing payments via the WorldPay platform."
    },
    {
      "name": "configuration",
      "description": "Operations available for configuring settings used by this API."
    }
  ],
  "paths": {
    "/payments/card/": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Requests a card payment be taken",
        "operationId": "createPayment",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "MakeCardPaymentRequest",
            "description": "Card payment request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MakeCardPaymentRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Payment successfully taken",
            "schema": {
              "$ref": "#/definitions/OrderResponse"
            }
          },
          "400": {
            "description": "Error Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/payments/{payment-provider-id}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Returns a payment by the given payment provider id",
        "operationId": "getPayment",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "payment-provider-id",
            "description": "Unique identifier that has been generated by the payment provider",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment details",
            "schema": {
              "$ref": "#/definitions/GetPaymentResponse"
            }
          },
          "400": {
            "description": "Error Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Payment not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/payments/api-key/": {
      "put": {
        "tags": [
          "configuration"
        ],
        "summary": "Sets the (payment provider issued) API key used by this service",
        "operationId": "api_key",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "UpdateApiKeyRequest",
            "description": "Payment provider API key",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateApiKeyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API Key updated",
            "schema": {
              "$ref": "#/definitions/UpdateApiKeyResponse"
            }
          },
          "400": {
            "description": "Serializer Error - Bad request"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "definitions": {
    "MakeCardPaymentRequest": {
      "type": "object",
      "required": [
        "card_holder_name",
        "card_number",
        "cvc",
        "expiry_month",
        "expiry_year",
        "currency_code",
        "customer_order_code",
        "order_description"
      ],
      "properties": {
        "amount": {
          "type": "integer",
          "example": 50000
        },
        "card_holder_name": {
          "type": "string",
          "example": "Mr Example Cardholder"
        },
        "card_number": {
          "type": "string",
          "example": 5454545454545454
        },
        "cvc": {
          "type": "string",
          "example": 352
        },
        "expiry_month": {
          "type": "integer",
          "example": 6
        },
        "expiry_year": {
          "type": "integer",
          "example": 2018
        },
        "currency_code": {
          "type": "string",
          "example": "GBP"
        },
        "customer_order_code": {
          "type": "string",
          "example": "OFS-MORE-162738"
        },
        "order_description": {
          "type": "string",
          "example": "Childminder Registration Fee"
        }
      }
    },
    "PaymentCreationResponse": {
      "type": "object",
      "required": [
        "paymentProviderOrderId"
      ],
      "properties": {
        "paymentProviderOrderId": {
          "type": "string",
          "format": "uuid",
          "example": "963e2f68-3d09-49e4-92e4-fd3869cc45fc"
        }
      }
    },
    "GetPaymentRequest": {
      "type": "object",
      "required": [
        "paymentProviderOrderId"
      ],
      "properties": {
        "paymentProviderOrderId": {
          "type": "string",
          "format": "uuid",
          "example": "963e2f68-3d09-49e4-92e4-fd3869cc45fc"
        }
      }
    },
    "OrderResponse": {
      "type": "object",
      "properties": {
        "riskScore": {
          "type": "object",
          "properties": {
            "value": {
              "type": "integer",
              "example": 1
            }
          }
        },
        "token": {
          "type": "string",
          "example": "TEST_SU_494d4873-c3d7-4dc6-b012-9925072195a9"
        },
        "amount": {
          "type": "string",
          "example": 50000
        },
        "paymentStatus": {
          "type": "string",
          "example": "SUCCESS"
        },
        "customerOrderCode": {
          "type": "string",
          "example": "TEST-CODE"
        },
        "orderCode": {
          "type": "string",
          "example": "a4c34c3c-cf1c-4404-a8fe-8b2d09b18e35"
        },
        "environment": {
          "type": "string",
          "example": "TEST"
        },
        "orderDescription": {
          "type": "string",
          "example": "Order description"
        },
        "currencyCode": {
          "type": "string",
          "example": "GBP"
        },
        "resultCodes": {
          "type": "object",
          "properties": {
            "avsResultCode": {
              "type": "string",
              "example": "NOT_SUPPLIED_BY_SHOPPER"
            },
            "cvcResultCode": {
              "type": "string",
              "example": "APPROVED"
            }
          }
        },
        "paymentResponse": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "ObfuscatedCard"
            },
            "expiryYear": {
              "type": "string",
              "example": 2020
            },
            "cardType": {
              "type": "string",
              "example": "MASTERCARD_DEBIT"
            },
            "name": {
              "type": "string",
              "example": "cardholder-name"
            },
            "cardSchemeType": {
              "type": "string",
              "example": "consumer"
            },
            "cardClass": {
              "type": "string",
              "example": "debit"
            },
            "countryCode": {
              "type": "string",
              "example": "GB"
            },
            "cardProductTypeDescNonContactless": {
              "type": "string",
              "example": "Dr MasterCard EEA"
            },
            "cardSchemeName": {
              "type": "string",
              "example": "MCI DEBIT"
            },
            "prepaid": {
              "type": "string",
              "example": "unknown"
            },
            "maskedCardNumber": {
              "type": "string",
              "example": "**** **** **** 5454"
            },
            "cardProductTypeDescContactless": {
              "type": "string",
              "example": "CL Dr MasterCard EEA"
            },
            "cardIssuer": {
              "type": "string",
              "example": "METRO BANK PLC"
            },
            "expiryMonth": {
              "type": "string",
              "example": 2
            }
          }
        }
      }
    },
    "GetPaymentResponse": {
      "type": "object",
      "properties": {
        "currencyCodeExponent": {
          "type": "integer",
          "example": 2
        },
        "creationDate": {
          "type": "string",
          "example": "2017-12-21T12:42:10.091+0000"
        },
        "modificationDate": {
          "type": "string",
          "example": "2017-12-21T12:42:10.141+0000"
        },
        "orderCode": {
          "type": "string",
          "example": "a4c34c3c-cf1c-4404-a8fe-8b2d09b18e35"
        },
        "token": {
          "type": "string",
          "example": "TEST_SU_494d4873-c3d7-4dc6-b012-9925072195a9"
        },
        "orderDescription": {
          "type": "string",
          "example": "Order description"
        },
        "amount": {
          "type": "string",
          "example": 50000
        },
        "currencyCode": {
          "type": "string",
          "example": "GBP"
        },
        "paymentStatus": {
          "type": "string",
          "example": "SUCCESS"
        },
        "paymentResponse": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "ObfuscatedCard"
            },
            "name": {
              "type": "string",
              "example": "cardholder-name"
            },
            "expiryMonth": {
              "type": "string",
              "example": 2
            },
            "expiryYear": {
              "type": "string",
              "example": 2020
            },
            "cardType": {
              "type": "string",
              "example": "MASTERCARD_DEBIT"
            },
            "maskedCardNumber": {
              "type": "string",
              "example": "**** **** **** 5454"
            },
            "cardSchemeType": {
              "type": "string",
              "example": "consumer"
            },
            "cardSchemeName": {
              "type": "string",
              "example": "MCI DEBIT"
            },
            "cardIssuer": {
              "type": "string",
              "example": "METRO BANK PLC"
            },
            "countryCode": {
              "type": "string",
              "example": "GB"
            },
            "cardClass": {
              "type": "string",
              "example": "debit"
            },
            "cardProductTypeDescNonContactless": {
              "type": "string",
              "example": "Dr MasterCard EEA"
            },
            "cardProductTypeDescContactless": {
              "type": "string",
              "example": "CL Dr MasterCard EEA"
            },
            "prepaid": {
              "type": "string",
              "example": "unknown"
            }
          }
        },
        "customerOrderCode": {
          "type": "string",
          "example": "TEST-CODE"
        },
        "environment": {
          "type": "string",
          "example": "TEST"
        },
        "riskScore": {
          "type": "object",
          "properties": {
            "value": {
              "type": "integer",
              "example": 1
            }
          }
        },
        "resultCodes": {
          "type": "object",
          "properties": {
            "avsResultCode": {
              "type": "string",
              "example": "NOT_SUPPLIED_BY_SHOPPER"
            },
            "cvcResultCode": {
              "type": "string",
              "example": "APPROVED"
            }
          }
        },
        "orderType": {
          "type": "string",
          "example": "ECOM"
        },
        "history": {
          "type": "object",
          "properties": {
            "modificationDate": {
              "type": "string",
              "example": "2017-12-21T12:42:10.151+0000"
            },
            "state": {
              "type": "string",
              "example": "SUCCESS"
            },
            "amount": {
              "type": "string",
              "example": 50000
            },
            "currencyCode": {
              "type": "string",
              "example": "GBP"
            },
            "currencyCodeExponent": {
              "type": "integer",
              "example": 2
            }
          }
        }
      }
    },
    "UpdateApiKeyRequest": {
      "type": "object",
      "required": [
        "api_key"
      ],
      "properties": {
        "api_key": {
          "type": "string",
          "example": "T_S_936c0508-5fae-4cb3-a631-24320e8ba123"
        }
      }
    },
    "UpdateApiKeyResponse": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}